requirements.txt
altair==5.4.1
altgraph==0.17.4
annotated-types==0.7.0
attrs==24.2.0
bertopic==0.16.3
blinker==1.8.2
blis==0.7.11
build==1.2.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
cymem==2.0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
funcy==2.0
fuzzywuzzy==0.18.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
hdbscan==0.8.38.post1
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
Levenshtein==0.26.0
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
narwhals==1.8.3
networkx==3.3
nltk==3.8.1
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pefile==2024.8.26
pillow==10.4.0
pip-tools==7.4.1
pipdeptree==2.23.4
plotly==5.24.1
preshed==3.0.9
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyLDAvis==3.4.1
pynndescent==0.5.13
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.2.0
PySimpleGUI==5.0.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-Levenshtein==0.26.0
python-pdf==0.40
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentence-transformers==3.1.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.38.0
sympy==1.13.3
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.6
urllib3==2.2.3
wasabi==1.1.3
watchdog==4.0.2
weasel==0.4.1
wheel==0.44.0
wrapt==1.16.0
